backend/requirements.txt
# FastAPI + scheduling
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
python-dotenv==1.0.1

# Azure CosmosDB
azure-cosmos==4.7.0
aiohttp

# ML dependencies
numpy
pandas
scikit-learn
xgboost
shap
lightgbm
joblib
